+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Added support for drag and drop tasks from task list to active page. #61
 - Added support to search/filter for tasks by closed date. #66
+- Added display of closed date on completed tasks in task list.
+- Added option to show closed date in embedded task lists.
 - Fixed task line not styled after code block with alternate delimiters in content.
 - Fixed preserve task line list markers when copying/moving tasks.
 - Fixed preserve task sub bullet content when copying/moving tasks.